Extract the response logic
